- added scaladoc to more classes in the database submodule
