+++ lending-database/src/main/java/info/armado/ausleihe/database/access/IdentityCardDAO.java
@@ -1,9 +1,9 @@
 package info.armado.ausleihe.database.access;
 
-import info.armado.ausleihe.database.dataobjects.Barcode;
-import info.armado.ausleihe.database.dataobjects.Envelope;
-import info.armado.ausleihe.database.dataobjects.Game;
-import info.armado.ausleihe.database.dataobjects.IdentityCard;
+import info.armado.ausleihe.database.barcode.Barcode;
+import info.armado.ausleihe.database.entities.Envelope;
+import info.armado.ausleihe.database.entities.Game;
+import info.armado.ausleihe.database.entities.IdentityCard;
 
 import java.util.List;
 import java.util.Set;
