+++ lending-rest-server/src/main/scala/info/armado/ausleihe/remote/IssueIdentityCards.scala
@@ -1,37 +1,18 @@
 package info.armado.ausleihe.remote
 
-import info.armado.ausleihe.database.dataobjects.ValidBarcode
-import info.armado.ausleihe.remote.results.IncorrectBarcode
-import javax.ws.rs.QueryParam
-import info.armado.ausleihe.database.access.IdentityCardDAO
-import info.armado.ausleihe.database.access.EnvelopeDAO
-import info.armado.ausleihe.database.access.LendIdentityCardDAO
-import info.armado.ausleihe.remote.results.LendingEntityNotExists
-import info.armado.ausleihe.remote.results.IssueIdentityCardSuccess
-import info.armado.ausleihe.util.DOExtensions._
-import info.armado.ausleihe.remote.results.LendingEntityInUse
-import info.armado.ausleihe.remote.dataobjects.inuse.IdentityCardInUse
-import info.armado.ausleihe.remote.dataobjects.inuse.EnvelopeInUse
-import javax.transaction.Transactional
-import javax.ws.rs.GET
-import javax.ws.rs.Path
-import javax.ws.rs.Produces
-import javax.ws.rs.core.MediaType
 import javax.enterprise.context.RequestScoped
 import javax.inject.Inject
-import info.armado.ausleihe.database.dataobjects.ValidateBarcode
-import info.armado.ausleihe.database.dataobjects.InvalidBarcode
-import info.armado.ausleihe.database.dataobjects.InvalidBarcode
-import info.armado.ausleihe.database.dataobjects.LendIdentityCard
-import javax.ws.rs.POST
-import javax.ws.rs.Consumes
+import javax.transaction.Transactional
+import javax.ws.rs._
+import javax.ws.rs.core.MediaType
+
+import info.armado.ausleihe.database.access.{EnvelopeDAO, IdentityCardDAO, LendIdentityCardDAO}
+import info.armado.ausleihe.database.barcode._
+import info.armado.ausleihe.database.entities.{Envelope, IdentityCard, LendIdentityCard}
+import info.armado.ausleihe.remote.dataobjects.inuse.{EnvelopeInUse, IdentityCardInUse}
 import info.armado.ausleihe.remote.requests.IssueIdentityCardRequest
-import info.armado.ausleihe.remote.results.AbstractResult
-import javax.ws.rs.core.Response
-import info.armado.ausleihe.database.dataobjects.IdentityCard
-import info.armado.ausleihe.database.dataobjects.Envelope
-import info.armado.ausleihe.database.dataobjects.Barcode
-import javax.ws.rs.BadRequestException
+import info.armado.ausleihe.remote.results._
+import info.armado.ausleihe.util.DOExtensions._
 
 @Path("/issue")
 @RequestScoped
