+++ lending-web-server/src/main/java/info/armado/ausleihe/faces/beans/GamesAdder.java
@@ -5,6 +5,9 @@ package info.armado.ausleihe.faces.beans;
 
 import info.armado.ausleihe.database.access.GamesDAO;
 import info.armado.ausleihe.database.dataobjects.*;
+import info.armado.ausleihe.database.barcode.Barcode;
+import info.armado.ausleihe.database.barcode.Barcode$;
+import info.armado.ausleihe.database.entities.Game;
 import info.armado.ausleihe.database.enums.Prefix;
 
 import java.util.ArrayList;
