+++ lending-web-server/src/main/java/info/armado/ausleihe/faces/beans/games/AddGameBean.java
@@ -14,7 +14,7 @@ import javax.inject.Named;
 import javax.transaction.Transactional;
 
 import info.armado.ausleihe.database.access.GamesDAO;
-import info.armado.ausleihe.database.dataobjects.Game;
+import info.armado.ausleihe.database.entities.Game;
 import info.armado.ausleihe.database.dataobjects.GameDuration;
 import info.armado.ausleihe.database.dataobjects.PlayerCount;
 import lombok.Getter;
