+++ lending-web-server/src/main/java/info/armado/ausleihe/faces/beans/games/ShowGameInfosBean.java
@@ -4,7 +4,7 @@
 package info.armado.ausleihe.faces.beans.games;
 
 import info.armado.ausleihe.database.access.GamesDAO;
-import info.armado.ausleihe.database.dataobjects.Game;
+import info.armado.ausleihe.database.entities.Game;
 import info.armado.ausleihe.database.objects.GameInfo;
 
 import java.io.Serializable;
