+++ lending-web-server/src/main/java/info/armado/ausleihe/faces/beans/identitycards/DeactivateIdentityCardBean.java
@@ -5,7 +5,7 @@ package info.armado.ausleihe.faces.beans.identitycards;
 
 import info.armado.ausleihe.database.access.IdentityCardDAO;
 import info.armado.ausleihe.database.access.LendIdentityCardDAO;
-import info.armado.ausleihe.database.dataobjects.IdentityCard;
+import info.armado.ausleihe.database.entities.IdentityCard;
 
 import java.io.Serializable;
 import java.util.List;
