+++ lending-web-server/src/main/java/info/armado/ausleihe/faces/converter/EnvelopeConverter.java
@@ -4,7 +4,9 @@
 package info.armado.ausleihe.faces.converter;
 
 import info.armado.ausleihe.database.access.EnvelopeDAO;
-import info.armado.ausleihe.database.dataobjects.*;
+import info.armado.ausleihe.database.barcode.Barcode;
+import info.armado.ausleihe.database.barcode.Barcode$;
+import info.armado.ausleihe.database.entities.Envelope;
 
 import javax.faces.application.FacesMessage;
 import javax.faces.component.UIComponent;
