+++ lending-web-server/src/main/java/info/armado/ausleihe/faces/converter/IdentityCardConverter.java
@@ -4,7 +4,9 @@
 package info.armado.ausleihe.faces.converter;
 
 import info.armado.ausleihe.database.access.IdentityCardDAO;
-import info.armado.ausleihe.database.dataobjects.*;
+import info.armado.ausleihe.database.barcode.Barcode;
+import info.armado.ausleihe.database.barcode.Barcode$;
+import info.armado.ausleihe.database.entities.IdentityCard;
 
 import javax.faces.application.FacesMessage;
 import javax.faces.component.UIComponent;
